+++ src/lang/enUS.lang
@@ -35,4 +35,4 @@ title:version=Version
 title:description=Description
 
 tab:mods=Manage Mods
-tab:search=Search GameBanana
+tab:search=Search Online
