+++ src/lang/jaJP.lang
@@ -32,4 +32,4 @@ title:version=バージョン
 title:description=説明
 
 tab:mods=mod管理
-tab:search=GameBanana検索
+tab:search=ネット検索
