+++ ui/widgets/search-table.slint
@@ -5,7 +5,7 @@ import { SearchPreviewData } from "../diva-types.slint";
 import { FontSettings as CosmicFontSettings } from "../styling.slint";
 import { GameBananaLogic, LangTL } from "../applogic.slint";
 
-export component GbSearchTable {
+export component OnlineSearchTable {
     in-out property <[SearchPreviewData]> results;
     in-out property <[TableColumn]> columns: [
         { title: "Preview", width: 118px },
